TitlePane: move files up one folder
