+++ src/main.ts
@@ -10,7 +10,7 @@ import Slider from './slider/Slider';
 import TabPane from './tabpane/TabPane';
 import TextInput from './textinput/TextInput';
 import Textarea from './textarea/Textarea';
-import TitlePane from './components/titlepane/TitlePane';
+import TitlePane from './titlepane/TitlePane';
 
 export {
 	Button,
