+++ src/titlepane/example/index.html
@@ -32,7 +32,7 @@
 			]
 		});
 
-		require([ 'src/examples/titlepane/index' ]);
+		require([ 'src/titlepane/example/index' ]);
 	</script>
 </body>
 </html>
