+++ src/titlepane/styles/titlePane.css
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.css';
+@import '../../styles/variables.css';
 
 .main {
 	overflow: hidden;
